tests: refactor agent tests for readability
